requirements.txt
aiogram==3.0.0b9
sqlalchemy==2.0.*
aiosqlite==0.19.*
python-dotenv==1.0.*